Removed testing main, fixed most active bug and Refactored Graph Tests
Removed the main in the compress-decompress.cpp that was made for testing purposes.
Fixed the Most Active function in the graph that it only counts the number of followers and doesn't count the number of following edges.
Modified the testing XML file for a more complex graph for testing.
Refactored the Graph Tests to suite the new graph